feat @LogicDelete

+++ src/main/java/com/github/wz2cool/dynamic/mybatis/mapper/provider/DynamicQueryProvider.java
@@ -1,18 +1,17 @@
 package com.github.wz2cool.dynamic.mybatis.mapper.provider;
 
-import com.github.wz2cool.dynamic.BaseFilterDescriptor;
 import com.github.wz2cool.dynamic.DynamicQuery;
-import com.github.wz2cool.dynamic.BaseSortDescriptor;
 import com.github.wz2cool.dynamic.UpdateQuery;
-import com.github.wz2cool.dynamic.mybatis.ParamExpression;
 import com.github.wz2cool.dynamic.mybatis.QueryHelper;
 import com.github.wz2cool.dynamic.mybatis.mapper.constant.MapperConstants;
-import com.github.wz2cool.dynamic.mybatis.mapper.helper.DynamicQuerySqlHelper;
 import com.github.wz2cool.dynamic.mybatis.mapper.helper.BaseEnhancedMapperTemplate;
+import com.github.wz2cool.dynamic.mybatis.mapper.helper.DynamicQuerySqlHelper;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.mapping.MappedStatement;
+import org.apache.ibatis.mapping.SqlCommandType;
 import tk.mybatis.mapper.mapperhelper.MapperHelper;
 import tk.mybatis.mapper.mapperhelper.SqlHelper;
+import tk.mybatis.mapper.util.MetaObjectUtil;
 
 import java.util.Map;
 
@@ -42,7 +41,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(SqlHelper.selectCount(entityClass));
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -52,7 +51,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(DynamicQuerySqlHelper.getSelectMax());
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -62,7 +61,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(DynamicQuerySqlHelper.getSelectMin());
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -72,7 +71,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(DynamicQuerySqlHelper.getSelectSum());
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -82,7 +81,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(DynamicQuerySqlHelper.getSelectAvg());
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -90,8 +89,16 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         Class<?> entityClass = getEntityClass(ms);
         StringBuilder sql = new StringBuilder();
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
-        sql.append(SqlHelper.deleteFromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        if (SqlHelper.hasLogicDeleteColumn(entityClass)) {
+            sql.append(SqlHelper.updateTable(entityClass, tableName(entityClass)));
+            sql.append("<set>");
+            sql.append(SqlHelper.logicDeleteColumnEqualsValue(entityClass, true));
+            sql.append("</set>");
+            MetaObjectUtil.forObject(ms).setValue("sqlCommandType", SqlCommandType.UPDATE);
+        } else {
+            sql.append(SqlHelper.deleteFromTable(entityClass, tableName(entityClass)));
+        }
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -106,7 +113,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         //支持查询指定列
         sql.append(DynamicQuerySqlHelper.getSelectColumnsClause());
         sql.append(SqlHelper.fromTable(entityClass, tableName(entityClass)));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         sql.append(DynamicQuerySqlHelper.getSortClause());
         return sql.toString();
     }
@@ -129,7 +136,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getUpdateBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(SqlHelper.updateTable(entityClass, tableName(entityClass), "example"));
         sql.append(DynamicQuerySqlHelper.getSetClause());
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
 
@@ -139,7 +146,7 @@ public class DynamicQueryProvider extends BaseEnhancedMapperTemplate {
         sql.append(DynamicQuerySqlHelper.getBindFilterParams(ms.getConfiguration().isMapUnderscoreToCamelCase()));
         sql.append(SqlHelper.updateTable(entityClass, tableName(entityClass), "example"));
         sql.append(SqlHelper.updateSetColumns(entityClass, "record", noNull, isNotEmpty()));
-        sql.append(DynamicQuerySqlHelper.getWhereClause());
+        sql.append(DynamicQuerySqlHelper.getWhereClause(entityClass));
         return sql.toString();
     }
